requirements.txt
click>=8.0.0
requests>=2.28.0
PyYAML>=6.0
colorama>=0.4.6
tqdm>=4.64.0
python-magic>=0.4.27
filetype>=1.2.0